Ship moving now always calculates movement error, removing additional checks after the movement.
